Fix notifications
