+++ server/libs/bot.js
@@ -26,8 +26,27 @@ To start receiving notifications please type:
 });
 
 const catPhoto = 'http://lorempixel.com/400/200/cats/'
-bot.command('cat', ({replyWithPhoto}) => replyWithPhoto(catPhoto))
+bot.command('cat', ({replyWithPhoto}) => replyWithPhoto(catPhoto));
 
+/**
+* Отправляет Telegram-оповещение пользователю
+*
+* @param {Object} notificationParams параметры нотификации
+*/
+bot.sendNotification = (notificationParams) => {
+    const {chatId} = notificationParams.user;
+    const {card, phone, amount} = notificationParams;
+    const cardNumberSecure = card.cardNumber.substr(card.cardNumber.length - 4);
+    var message;
+    if (notificationParams.type === 'paymentMobile') {
+        message = `С вашей 💳  **** **** **** ${cardNumberSecure} было переведено ${amount}${card.currency} на 📱 ${phone}`;
+    } else {
+        message = `На вашу 💳  **** **** **** ${cardNumberSecure} поступило ${amount}${card.currency}`;
+    }
+    if (chatId) {
+        bot.telegram.sendMessage(chatId, message);
+    }
+}
 
 // Start polling
 bot.startPolling()
